third_party/blink/web_tests/fast/forms/week-multiple-fields/week-multiple-fields-keyboard-events-expected.txt

Multiple fields UI of week input type with keyboard events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Please run this with DumpRenderTree.

Test following keys:
Digits
Left/Right - Move focus field inside element
Up/Down - Increment/decrement value of focus field
Tab - Move focus field
Backspace - Make value empty
  
== Digit keys ==
PASS input.value is "0012-W09"
== Digit keys starting with zero ==
PASS input.value is "0300-W02"
== Digit keys and backspace key ==
PASS input.value is "0006-W01"
== Left/Right keys ==
PASS input.value is "0005-W06"
PASS document.activeElement.id is "input"
== Up/Down keys ==
PASS input.value is "2012-W10"
PASS input.value is "2012-W08"
== Up/Down keys on empty value ==
PASS input.value is currentYear + "-W52"
== Up/Down keys on empty value 2 ==
PASS input.value is currentYear + "-W02"
== Tab key ==
PASS input.value is "0005-W09"
PASS input.value is "0005-W07"
PASS document.activeElement.id is "after"
== Shift+Tab key ==
PASS input.value is "0003-W09"
PASS document.activeElement.id is "before"
== Up key on maximum value ==
PASS input.value is "0001-W10"
== Up key with a maximum attribute ==
PASS input.value is "1000-W13"
== Down key on minimum value ==
PASS input.value is ""
== Down key with a minimum attribute ==
PASS input.value is ""
== Inconsistent min-max attributes ==
PASS input.value is "1000-W13"
PASS input.value is "1999-W13"
== Make an invalid date ==
PASS input.value is ""
== Backspace key ==
PASS input.value is ""
== Delete key ==
PASS input.value is ""
== Typeahead ==
PASS input.value is "0001-W12"
PASS input.value is "0002-W12"
== Disabled/readonly ==
PASS input.value is "2012-W10"
PASS input.value is "2012-W11"
PASS input.value is "2012-W11"
PASS input.value is "2012-W12"
PASS successfullyParsed is true

TEST COMPLETE
